+++ packages/orq-rc/src/models/components/evaluatorresponsehttp.ts
@@ -210,8 +210,8 @@ export const EvaluatorResponseHttp$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-06-29T13:49:50.653Z"),
-  updated: z.string().default("2026-06-29T13:49:50.653Z"),
+  created: z.string().default("2026-06-29T16:31:24.066Z"),
+  updated: z.string().default("2026-06-29T16:31:24.066Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.nullable(
     z.union([