+++ packages/orq-rc/RELEASES.md
@@ -10339,3 +10339,13 @@ Based on:
 - [typescript v4.11.0-rc.55] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.55 - packages/orq-rc
+
+## 2026-06-29 16:34:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.56] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.56 - packages/orq-rc